+++ src/api/canvas/newQuiz/newQuizItemsApi.ts
@@ -12,7 +12,7 @@ import {
   transformToCanvasNewQuizOrderingItem,
   transformToCanvasNewQuizMultiAnswerItem,
   transformToCanvasNewQuizMatchingItem,
-} from "../../../helper/transformForCanvasNewQuiz.js";
+} from "../../../utils/transformForCanvasNewQuiz.js";
 
 const baseUrl = process.env.BASE_URL;
 const apiToken = process.env.API_TOKEN;
