+++ src/menus/canvasCourseActions/handleDeleteNewQuiz.ts
